refactor: move document visualiation outside variables tab
