+++ operate/client/src/App/ProcessInstance/DocumentsView/DocumentPreview/PreviewImage.tsx
