+++ operate/client/src/App/ProcessInstance/DocumentsView/DocumentPreview/styled.tsx
