+++ operate/client/src/App/ProcessInstance/DocumentsView/documentInfo.ts
@@ -7,8 +7,6 @@
  */
 
 import {z} from 'zod';
-import {safeJsonParse} from 'modules/utils';
-import {untruncateJson} from 'modules/utils/editor/untruncateJSON';
 import {mergePathname} from 'modules/request/mergePathname';
 import {getClientConfig} from 'modules/utils/getClientConfig';
 import {
@@ -38,14 +36,6 @@ type DocumentInfo = {
   size: number;
   isExpired: boolean;
 };
-
-type DocumentParseResult =
-  | {type: 'single'; document: DocumentInfo}
-  | {type: 'list'; documents: DocumentInfo[]; isLowerBound: boolean};
-
-function isDocumentReference(value: unknown): value is DocumentReference {
-  return relaxedDocumentReferenceSchema.safeParse(value).success;
-}
 
 const MIME_TYPE_MAP: Record<string, DocumentType> = {
   'image/jpeg': 'image',
@@ -87,72 +77,5 @@ function toDocumentInfo(ref: DocumentReference): DocumentInfo {
   };
 }
 
-function parseDocumentVariable(
-  value: string,
-  isTruncated: boolean,
-): DocumentParseResult | null {
-  const {completed, collectionDepth} = isTruncated
-    ? untruncateJson(value)
-    : {completed: value, collectionDepth: 0};
-
-  const parsed = safeJsonParse(completed);
-
-  if (isDocumentReference(parsed)) {
-    return {type: 'single', document: toDocumentInfo(parsed)};
-  }
-
-  if (!Array.isArray(parsed) || parsed.length === 0) {
-    return null;
-  }
-
-  const docs = parsed.filter(isDocumentReference);
-  if (docs.length === 0) {
-    return null;
-  }
-
-  if (!isTruncated && docs.length !== parsed.length) {
-    return null;
-  }
-
-  if (docs.length === 1 && !isTruncated) {
-    return {type: 'single', document: toDocumentInfo(docs[0]!)};
-  }
-
-  return {
-    type: 'list',
-    documents: docs.map(toDocumentInfo),
-    isLowerBound: collectionDepth > 0 || docs.length < parsed.length,
-  };
-}
-
-const UNITS = ['B', 'KiB', 'MiB', 'GiB'];
-const BYTES_BASE = 1024;
-
-/**
- * Formats a number of bytes into a human readable string with binary prefixes (up to GiB)
- * @param bytes - The number of bytes to format
- * @returns Formatted string (e.g. "1.5 MiB")
- */
-function toHumanReadableBytes(bytes: number): string {
-  if (bytes === 0) {
-    return '0 B';
-  }
-
-  if (!Number.isFinite(bytes)) {
-    return 'N/A';
-  }
-
-  const exponent = Math.min(
-    Math.floor(Math.log(bytes) / Math.log(BYTES_BASE)),
-    UNITS.length - 1,
-  );
-
-  const value = bytes / Math.pow(BYTES_BASE, exponent);
-  const unit = UNITS[exponent];
-  const formatted = value.toFixed(2).replace(/\.?0+$/, '');
-
-  return `${formatted} ${unit}`;
-}
-
-export {parseDocumentVariable, toHumanReadableBytes};
-export type {DocumentInfo, DocumentParseResult};
+export {relaxedDocumentReferenceSchema, toDocumentInfo};
+export type {DocumentInfo, DocumentReference};
