+++ operate/client/src/App/ProcessInstance/DocumentsView/middleTruncate.ts
